+++ src/Sql/Sql.php
@@ -277,7 +277,7 @@ class Sql extends SqlLexer implements Interface\Parser
             'RANDOM_BYTES' => $query->randomBytes((int) ($arguments[0] ?? 10)),
             'ARRAY_COMBINE' => $query->arrayCombine((string) ($arguments[0] ?? ''), (string) ($arguments[1] ?? '')),
             'ARRAY_MERGE' => $query->arrayMerge((string) ($arguments[0] ?? ''), (string) ($arguments[1] ?? '')),
-            'DATE_FORMAT' => $query->formatDate((string) ($arguments[0] ?? ''), (string) ($arguments[1] ?? '')),
+            'DATE_FORMAT' => $query->formatDate((string) ($arguments[0] ?? ''), (string) ($arguments[1] ?? 'c')),
             'MATCH' => $this->processMatchFunction($query, $arguments),
             default => throw new Exception\UnexpectedValueException("Unknown function: $functionName"),
         };